services/rtVideoAnalysis/requirements.txt
opencv-python==4.8.1.78
azure-cognitiveservices-vision-computervision==0.9.0
kafka-python==2.0.2
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
numpy==1.24.3
schedule==1.2.0
websocket-client==1.6.4
stomp.py==8.2.0
